src/main/java/com/chainsys/bloodbankapp/model/BloodGroup.java
package com.chainsys.bloodbankapp.model;

public class BloodGroup {

	private int bloodGroupId;
	private String bloodGroup;

	@Override
	public String toString() {
		return "BloodGroup [bloodGroupId=" + bloodGroupId + ", bloodGroup=" + bloodGroup + "]";
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((bloodGroup == null) ? 0 : bloodGroup.hashCode());
		result = prime * result + bloodGroupId;
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		BloodGroup other = (BloodGroup) obj;
		if (bloodGroup == null) {
			if (other.bloodGroup != null)
				return false;
		} else if (!bloodGroup.equals(other.bloodGroup))
			return false;
		if (bloodGroupId != other.bloodGroupId)
			return false;
		return true;
	}

	public int getBloodGroupId() {
		return bloodGroupId;
	}

	public void setBloodGroupId(int bloodGroupId) {
		this.bloodGroupId = bloodGroupId;
	}

	public String getBloodGroup() {
		return bloodGroup;
	}

	public void setBloodGroup(String bloodGroup) {
		this.bloodGroup = bloodGroup;
	}

}
